# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Workshop API",
        "description": "API documentation for SSB Workshop application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Default Server"
        }
    ],
    "paths": {
        "/api/units": {
            "get": {
                "tags": [
                    "Equipment"
                ],
                "summary": "Get all units",
                "operationId": "af04c7ee3a7406286c567017520373f7",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "uid": {
                                                "type": "string"
                                            },
                                            "code": {
                                                "type": "string"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "engine_no": {
                                                "type": "string"
                                            },
                                            "chassis_no": {
                                                "type": "string"
                                            },
                                            "plate_number": {
                                                "type": "string"
                                            },
                                            "pic_unit": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/units/{id}": {
            "get": {
                "tags": [
                    "Equipment"
                ],
                "summary": "Find unit by ID",
                "operationId": "7f09ad6ced9de2f57a7bcf3010e21609",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Unit ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "uid": {
                                            "type": "string"
                                        },
                                        "code": {
                                            "type": "string"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "engine_no": {
                                            "type": "string"
                                        },
                                        "chassis_no": {
                                            "type": "string"
                                        },
                                        "plate_number": {
                                            "type": "string"
                                        },
                                        "pic_unit": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unit not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/units/uid/{uid}": {
            "get": {
                "tags": [
                    "Equipment"
                ],
                "summary": "Find unit by UID",
                "operationId": "e4876a135d3ca3f5679f897789e83a6c",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "Unit UID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "uid": {
                                            "type": "string"
                                        },
                                        "code": {
                                            "type": "string"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "engine_no": {
                                            "type": "string"
                                        },
                                        "chassis_no": {
                                            "type": "string"
                                        },
                                        "plate_number": {
                                            "type": "string"
                                        },
                                        "pic_unit": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unit not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Equipment",
            "description": "Equipment"
        }
    ]
}